feat(components): split new
