+++ src/components/index.tsx
@@ -1,3 +1,5 @@
 export * from './Checkmark'
+export * from './Menu'
 export * from './ModeSwitcher'
 export * from './PreferenceCheckbox'
+export * from './WelcomeScreen'
